Move Signal to utils.
